fix disable resetting color+temp continuously
To reproduce without this patch:

1. run `./src/redshift -v` in a terminal
2. in another terminal, `pkill -USR1 redshift`
   ... which ought to disable its processing altogether, once completed
   ... but doesn't :(
3. while it's disabled, manually set the brightness:
    `xrandr --output HDMI2 --brightness 0.5`
4. after a few seconds (default is 5secs between iterations), the
   brightness (and temp) jumps up again despite redshift being disabled

With this patch, instead - the resetting of colour and temperature is
done only once after the program's been disabled, and subsequent changes
of brightness or temperature etc. via a separate `xrandr` while the
program is disabled have the desired effect (i.e. they stick instead of
being reset by a "disabled" redshift program).

Once the program is re-enabled again (i.e. via another `pkill -USR1
redshift`), the program is again allowed to make changes to brightness
and temperature, as it ought to be able to.