fix(systemd): use graphical-session.target
The previous `After=` reference (`display-manager.service`) is not something
that's included with systemd. While it may be present on some user's
system (because they created it, or some specific package includes it),
it can't be expected to be present.

On the other hand, `graphical-session.target` is included with systemd,
and can be expected to be there, and be triggered when starting a
graphical session.

This guarantees that redshift will only run AFTER a graphic session is
up (eg: either Xorg or Wayland are up).

The previous `WantedBy=` value (`default.target`) meant that redshift got
started when a user session began, at which point we can't expect a graphic
session to be up and running (eg: Xorg or Wayland may have not even been
invoked yet).

Setting `WantedBy=graphical-session.target` ensures that, when the
service is enable, it'll only get run when starting a graphical session
(again, either Xorg or Wayland, whatever DE's may be used).

Add a FAQ entry on issues with running via systemd

Fixes #742