fix(indicator): handle log prefix while parsing
This is related to the recent log changes: 6e785e65f074dd74889b06901c3a314d98411122

The log format changed in two ways:
- Each entry of the log is prefixed by the loglevel
    + *Before*: `Something: Value`
    + *After*: `Loglevel: Something: Value`
- All log output is now in stderr instead of a combination of stdout
  (for info messages) and stderr (for error messages)

However, `gammastep-indicator` needs to parse the log to work, and since
no modification was done in it the output was broken, specially in
"Enable" and "Info" items. This commit fixes it.

Co-authored-by: Cameron Nemo <cnemo@tutanota.com>

closes #21