fix(indicator): escape error message in dialog
The error message popup was showing empty with the following error
message showing in the stderr:

```
</i>' from markup due to error parsing markup: Error on line 2 char 45: “ ” is not a valid character following a “<” character; it may not begin an element name
```

It happening because the error message sometimes contain either < or >
characters, the XML markup was trying to interpret as a tag and failing.

This commit fixes it by escaping the text to be XML safe.